Updated based on core PR
